requirements.txt
annotated-types==0.7.0
anyio==4.8.0
argostranslate==1.9.6
audioread==3.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.10
contourpy==1.3.1
ctranslate2==4.5.0
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
filelock==3.17.0
flet==0.27.1
fonttools==4.56.0
fsspec==2025.2.0
googletrans==4.0.2
gTTS==2.5.4
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.0
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
noisereduce==3.0.3
numba==0.61.0
numpy==2.1.3
oauthlib==3.2.2
openai==1.63.2
openai-whisper==20240930
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.29.3
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.6.1
pyparsing==3.2.1
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.98
pywin32==308
PyYAML==6.0.2
regex==2024.11.6
repath==0.9.0
requests==2.32.3
sacremoses==0.0.53
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soxr==0.5.0.post1
SpeechRecognition==3.14.1
stanza==1.1.1
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
urllib3==2.3.0
